Separate styles into multiple files
