+++ src/gui/styles/am2r_style.py
@@ -1,0 +1,26 @@
+from gui.styles.base_style import BaseStyle
+
+
+class AM2RStyle(BaseStyle):
+    background = {
+        "Distribution Center": "#9fc5e8",
+        "GFS Thoth": "#f9944a",
+        "Golden Temple": "#ffe599",
+        "Hydro Station": "#b6d7a8",
+        "Industrial Complex": "#b4a7d6",
+        "Main Caves": "#b05900",
+        "The Depths": "#38761d",
+        "The Tower": "#dd7e6b",
+    }
+    foreground = {
+        "Distribution Center": "black",
+        "GFS Thoth": "black",
+        "Golden Temple": "black",
+        "Hydro Station": "black",
+        "Industrial Complex": "black",
+        "Main Caves": "white",
+        "The Depths": "white",
+        "The Tower": "black",
+    }
+
+    victory_background = "#0ce561"
