+++ src/gui/styles/base_style.py
@@ -1,0 +1,6 @@
+class BaseStyle:
+    background: dict[str, str] = dict()
+    foreground: dict[str, str] = dict()
+    fallback_background: str = "black"
+    fallback_foreground: str = "white"
+    victory_background: str = "#0ce561"
