+++ src/gui/styles/dread_style.py
@@ -1,0 +1,27 @@
+from gui.styles.base_style import BaseStyle
+
+
+class DreadStyle(BaseStyle):
+    background = {
+        "Artaria": "#ffe599",
+        "Burenia": "#72bddd",
+        "Cataris": "#dd7e6b",
+        "Dairon": "#eb9aca",
+        "Elun": "#b4a7d6",
+        "Ferenia": "#6d9eeb",
+        "Ghavoran": "#b6d7a8",
+        "Hanubia": "#a88160",
+    }
+
+    foreground = {
+        "Artaria": "black",
+        "Burenia": "black",
+        "Cataris": "black",
+        "Dairon": "black",
+        "Elun": "black",
+        "Ferenia": "black",
+        "Ghavoran": "black",
+        "Hanubia": "black",
+    }
+
+    victory_background = "#0ce561"
