+++ src/gui/styles/prime1_style.py
@@ -1,0 +1,21 @@
+from gui.styles.base_style import BaseStyle
+
+
+class Prime1Style(BaseStyle):
+    background = {
+        "Chozo Ruins": "#d8c4af",
+        "Magmoor Caverns": "#dd7e6b",
+        "Phazon Mines": "#b4a7d6",
+        "Phendrana Drifts": "#bfbfbf",
+        "Tallon Overworld": "#9fbc93",
+    }
+
+    foreground = {
+        "Chozo Ruins": "black",
+        "Magmoor Caverns": "black",
+        "Phazon Mines": "black",
+        "Phendrana Drifts": "black",
+        "Tallon Overworld": "black",
+    }
+
+    victory_background = "#17baba"
