+++ src/gui/styles/prime3_style.py
@@ -1,0 +1,39 @@
+from gui.styles.base_style import BaseStyle
+
+
+class Prime3Style(BaseStyle):
+    background = {
+        "Bryyo - Fire": "#f9944a",
+        "Bryyo - Ice": "#cfe2f3",
+        "Bryyo - Reptilicus": "#9fbc93",
+        "Bryyo - Seed": "#3c78d8",
+        "G.F.S. Olympus": "#8eafce",
+        "G.F.S. Valhalla": "#d5a6bd",
+        "Norion": "#bfbfbf",
+        "Pirate Homeworld - Command": "#dd7e6b",
+        "Pirate Homeworld - Mines": "#4dc923",
+        "Pirate Homeworld - Research": "#cc4125",
+        "Pirate Homeworld - Seed": "#3c78d8",
+        "SkyTown, Elysia - Main": "#d8c4af",
+        "SkyTown, Elysia - Pod": "#f9cb9c",
+        "SkyTown, Elysia - Seed": "#3c78d8",
+    }
+
+    foreground = {
+        "Bryyo - Fire": "black",
+        "Bryyo - Ice": "black",
+        "Bryyo - Reptilicus": "black",
+        "Bryyo - Seed": "white",
+        "G.F.S. Olympus": "black",
+        "G.F.S. Valhalla": "black",
+        "Norion": "black",
+        "Pirate Homeworld - Command": "black",
+        "Pirate Homeworld - Mines": "black",
+        "Pirate Homeworld - Research": "white",
+        "Pirate Homeworld - Seed": "white",
+        "SkyTown, Elysia - Main": "black",
+        "SkyTown, Elysia - Pod": "black",
+        "SkyTown, Elysia - Seed": "white",
+    }
+
+    victory_background = "#0ce561"
